usePrevious to ts
