+++ src/usePrevious.ts
@@ -3,8 +3,8 @@ import { getOrInvoke } from './utils'
 
 // Taken from React official docs (2019-06-14)
 // https://reactjs.org/docs/hooks-faq.html#how-to-get-the-previous-props-or-state
-export default function usePrevious(valueOrFactory) {
-  const ref = useRef()
+export default function usePrevious<T>(valueOrFactory: T | (() => T)): T {
+  const ref = useRef<T>()
   useEffect(() => {
     ref.current = getOrInvoke(valueOrFactory)
   })
